requirements.txt
pdfplumber
openai>=1.40.0
requests
pillow